started refactoring record provenance transformer
