+++ src/Makefile.am
@@ -129,6 +129,7 @@ souffle_sources = \
               WriteStreamCSV.h                          \
               IOSystem.h                                \
 			  RamInterface.cpp		RamInterface.h		\
+			  ProvenanceTransformer.cpp	ProvenanceTransformer.h	\
               Global.cpp            Global.h
 
 # -- build souffle as a library so it can be reused in testing
